:green_heart: Fix CI (at least in local)
Refactored the utility-db connection to count open connections. Now there is a custom
setting, rather than including it into the DATABASES setting. This ensures that pytest-django
does not run migrations for the postgres alias, which crashes on an existing
data migration as that uses the default alias (and there is a mismatch between
local model state in migrations and the actual database state).