:bug: Fix slug field migration with appropriate max_length
Previously, the migration that introduced the slug field to
the Service model used the default max_length of 50, which
conflicted with our data migration which set the field by
slugifying the api_root (which has a max_length of 255).

This commit addresses this by increasing the slug field size,
and also by checking for the limit in the migration. We had
to yank the 0.35.0 release because this migration would have
failed on a large proportion of deploys, so this will
be released as part of 0.35.1.