:children_crossing: Fix slug auto-populate
Auto-populate from the label instead of the api root, as the latter
field is located *below* the slug field and not intuitive.

I believe this stems from the time we didn't have explicit labels for
services or they weren't required. The data migration being based on
the api root is okay - it helps enforce uniqueness.