:sparkles: [open-formulieren/open-forms#5193] Ensure that the JWT claims have exp set
The exp claim is a NumericDate (number of seconds sinds epoch in UTC)
on or after which the token may no longer be processed. We calculate
it by taking the iat claim (issued at timestamp) and adding the number
of seconds configured on the service that a token remains valid.

This gives user control over their security hardening by allowing them
to use shorter or longer expiration times, and gives flexibility to
comply with different vendor requirements.